feat(auto): migrate autoTaskRunner to TypeScript
